+++ creator-node/src/components/tracks/tracksComponentService.js
@@ -19,11 +19,10 @@ const ENABLE_IPFS_ADD_TRACKS = config.get('enableIPFSAddTracks')
  *
  * @param {Object} logContext the context of the request used to create a generic logger
  * @param {Object} requestProps more request specific context, NOT the req object from Express
- * @param {Object} ipfsLatest ipfsLatest instance
  * @returns a success or error server response
  * @dev - Prune upload artifacts after successful and failed uploads. Make call without awaiting, and let async queue clean up.
  */
-const handleTrackContentRoute = async ({ logContext }, requestProps, ipfsLatest) => {
+const handleTrackContentRoute = async ({ logContext }, requestProps) => {
   const logger = genericLogger.child(logContext)
 
   const routeTimeStart = Date.now()
@@ -57,7 +56,6 @@ const handleTrackContentRoute = async ({ logContext }, requestProps, ipfsLatest)
     { logContext: requestProps.logContext },
     requestProps.session.cnodeUserUUID,
     transcodedFilePath,
-    ipfsLatest,
     ENABLE_IPFS_ADD_TRACKS
   )
 
@@ -71,7 +69,6 @@ const handleTrackContentRoute = async ({ logContext }, requestProps, ipfsLatest)
         { logContext: requestProps.logContext },
         requestProps.session.cnodeUserUUID,
         segmentAbsolutePath,
-        ipfsLatest,
         ENABLE_IPFS_ADD_TRACKS
       )
       return { multihash, srcPath: segmentFilePath, dstPath }
