+++ creator-node/src/utils.js
@@ -13,7 +13,7 @@ const redis = require('./redis')
 const config = require('./config')
 const BlacklistManager = require('./blacklistManager')
 const { generateTimestampAndSignature } = require('./apiSigning')
-const { ipfsAddNonImages } = require('./ipfsClient')
+const { ipfsAddNonImages } = require('./ipfsAdd')
 
 const readFile = promisify(fs.readFile)
 
@@ -227,7 +227,7 @@ const ipfsGet = ({ ipfsLatest }, logger, path, timeout = 1000) => new Promise(as
   }
 })
 
-async function findCIDInNetwork (ipfsLatest, filePath, cid, logger, libs, trackId = null) {
+async function findCIDInNetwork (filePath, cid, logger, libs, trackId = null) {
   const attemptedStateFix = await getIfAttemptedStateFix(filePath)
   if (attemptedStateFix) return
 
@@ -260,7 +260,7 @@ async function findCIDInNetwork (ipfsLatest, filePath, cid, logger, libs, trackI
         await writeStreamToFileSystem(resp.data, filePath, /* createDir */ true)
 
         // verify that the file written matches the hash expected if added to ipfs
-        const ipfsHashOnly = await ipfsAddNonImages(ipfsLatest, filePath, { onlyHash: true, timeout: 2000 })
+        const ipfsHashOnly = await ipfsAddNonImages(filePath, { onlyHash: true, timeout: 2000 })
 
         if (cid !== ipfsHashOnly) {
           await fs.unlink(filePath)
