+++ creator-node/test/fileManager.test.js
@@ -64,7 +64,7 @@ describe('test fileManager', () => {
      */
     it('should throw error if cnodeUserUUID is not present', async () => {
       try {
-        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, null, srcPath, ipfsLatest)
+        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, null, srcPath)
         assert.fail('Should not have passed if cnodeUserUUID is not present in request.')
       } catch (e) {
         assert.deepStrictEqual(e.message, 'User must be authenticated to save a file')
@@ -80,7 +80,7 @@ describe('test fileManager', () => {
       sinon.stub(ipfs, 'addFromFs').rejects(new Error('ipfs is down!'))
 
       try {
-        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath, ipfsLatest, true /* enableIPFSAdd */)
+        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath, true /* enableIPFSAdd */)
       } catch (e) {
         assert.fail('Should have passed if ipfs is down.')
       }
@@ -100,7 +100,7 @@ describe('test fileManager', () => {
       )
 
       try {
-        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath, ipfsLatest)
+        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath)
         assert.fail('Should not have passed if file copying fails.')
       } catch (e) {
         assert.deepStrictEqual(e.message, 'Failed to copy files!!')
@@ -119,7 +119,7 @@ describe('test fileManager', () => {
       sinon.stub(models.File, 'create').returns({ dataValues: { fileUUID: 'uuid' } })
 
       try {
-        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath, ipfsLatest, true /* enableIPFSAdd */)
+        await saveFileToIPFSFromFS({ logContext: { requestID: uuid() } }, req.session.cnodeUserUUID, srcPath, true /* enableIPFSAdd */)
       } catch (e) {
         assert.fail(e.message)
       }
