+++ creator-node/test/lib/helpers.js
@@ -6,7 +6,7 @@ const DiskManager = require('../../src/diskManager')
 
 const { handleTrackContentRoute } = require('../../src/components/tracks/tracksComponentService')
 
-const uploadTrack = async (filePath, cnodeUserUUID, ipfs, blacklistManager) => {
+const uploadTrack = async (filePath, cnodeUserUUID, blacklistManager) => {
   const { fileUUID, fileDir } = saveFileToStorage(filePath)
   const resp = await handleTrackContentRoute({
     logContext: {
@@ -24,7 +24,6 @@ const uploadTrack = async (filePath, cnodeUserUUID, ipfs, blacklistManager) => {
       cnodeUserUUID
     }
   },
-  ipfs,
   blacklistManager
   )
